Remove old method of storing giant ai data xml
See a2dff39 for the replacement code. This is now removed because the results appear the same.

It's a single commit, though, so it can be reverted if necessary.

Note that we no longer change the case of Selected Works or CV, since the xsl was updated to use the incoming title case. Time will tell if this has a negative effect.