Limit the 'EntityKeys' returned in the ai/digitalmeasures single person api call
This can reduce the response time by up to 50%, and results in a much smaller raw ai xml file for later processing.

The hope is that this will reduce memory usage and allow the script to run on platform.

The entity keys were chosen by scanning the xsl file for usage.